# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whatsapp/send:
    post:
      operationId: sendWhatsAppQuickMessage
      tags: [whatsapp]
      summary: Send a free-form WhatsApp message to a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsAppSendInput"
      responses:
        "200":
          description: Message queued / sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsAppSendResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
        "503":
          description: WhatsApp not connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }

  # ─── Webhook Ingestion ───────────────────────────────────────────
  /webhook/ingest:
    post:
      operationId: ingestWebhook
      tags: [webhook]
      summary: Receive event payloads from Team AI Embedded and other connected platforms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Ingestion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookIngestResult"
        "400":
          description: Validation failed
components:
...
  schemas:
...
    WebhookPayload:
      type: object
      required: [event, sourcePlatform]
      properties:
        event: { type: string }
        sourcePlatform: { type: string }
        taskData:
          type: object
          properties:
            title: { type: string }
            description: { type: string }
            assigneeId:
              type: ["integer", "null"]
            projectId: { type: integer }
            status: { type: string }
            priority: { type: string }
            dueDate:
              type: ["string", "null"]
            resourceRequired: {}
            deliveryFormat: { type: string }
            notifyVia: { type: string }
        meta: {}
    WebhookIngestResult:
      type: object
      required: [accepted, taskId, message]
      properties:
        accepted: { type: boolean }
        taskId:
          type: ["integer", "null"]
        message: { type: string }
        validationErrors:
          type: array
          items: { type: string }

    # ─── Team Auth Schemas ────────────────────────────────────────
...
    WhatsAppSendInput:
      type: object
      required: [to, message]
      properties:
        to:
          type: string
          minLength: 1
          description: Recipient phone number (e.g. +254712345678)
        message:
          type: string
          minLength: 1
          description: Plain-text message body
    WhatsAppSendResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        to: { type: string }